+++ persistence/src/main/scala/zio/actors/persistence/journal/InMemJournal.scala
@@ -5,6 +5,24 @@ import zio.actors.persistence.PersistenceId.PersistenceId
 import zio.actors.persistence.PersistenceConfig
 import InMemJournal.JournalRow
 
+private[actors] final class InMemJournal[Ev](journalRef: Ref[List[JournalRow[Ev]]]) extends Journal[Ev] {
+
+  override def persistEvent(persistenceId: PersistenceId, event: Ev): Task[Unit] =
+    for {
+      journal <- journalRef.get
+      maxSeq  = journal.collect { case row if row.persistenceId == persistenceId => row.seqNum }
+      max     = if (maxSeq.isEmpty) 0 else maxSeq.max
+      _       <- journalRef.set(journal :+ JournalRow(persistenceId, max + 1, event))
+    } yield ()
+
+  override def getEvents(persistenceId: PersistenceId): Task[Seq[Ev]] =
+    for {
+      journal <- journalRef.get
+      events  = journal.filter(_.persistenceId == persistenceId).sortBy(_.seqNum).map(_.event)
+    } yield events
+
+}
+
 private[actors] object InMemJournal {
 
   private case class JournalRow[Ev](persistenceId: PersistenceId, seqNum: Int, event: Ev)
@@ -37,23 +55,5 @@ private[actors] object InMemJournal {
 
   def make[Ev](): UIO[InMemJournal[Ev]] =
     Ref.make(List.empty[JournalRow[Ev]]).map(new InMemJournal[Ev](_))
-
-}
-
-private[actors] class InMemJournal[Ev](journalRef: Ref[List[JournalRow[Ev]]]) extends Journal[Ev] {
-
-  override def persistEvent(persistenceId: PersistenceId, event: Ev): Task[Unit] =
-    for {
-      journal <- journalRef.get
-      maxSeq  = journal.collect { case row if row.persistenceId == persistenceId => row.seqNum }
-      max     = if (maxSeq.isEmpty) 0 else maxSeq.max
-      _       <- journalRef.set(journal :+ JournalRow(persistenceId, max + 1, event))
-    } yield ()
-
-  override def getEvents(persistenceId: PersistenceId): Task[Seq[Ev]] =
-    for {
-      journal <- journalRef.get
-      events  = journal.filter(_.persistenceId == persistenceId).sortBy(_.seqNum).map(_.event)
-    } yield events
 
 }
